add multithreading for checks and split code into modules
